change integer fields in PSAPI_WORKING_SET_EX_BLOCK from int to uint64
Otherwise, they won't work in 386 machines.

+++ windows/syscall_windows.go
@@ -1718,13 +1718,13 @@ func (b PSAPI_WORKING_SET_EX_BLOCK) Valid() bool {
 }
 
 // The number of processes that share this page. The maximum value of this member is 7.
-func (b PSAPI_WORKING_SET_EX_BLOCK) ShareCount() int {
+func (b PSAPI_WORKING_SET_EX_BLOCK) ShareCount() uint64 {
 	return b.intField(1, 3)
 }
 
 // The memory protection attributes of the page. For a list of values, see
 // https://docs.microsoft.com/en-us/windows/win32/memory/memory-protection-constants
-func (b PSAPI_WORKING_SET_EX_BLOCK) Win32Protection() int {
+func (b PSAPI_WORKING_SET_EX_BLOCK) Win32Protection() uint64 {
 	return b.intField(4, 11)
 }
 
@@ -1734,7 +1734,7 @@ func (b PSAPI_WORKING_SET_EX_BLOCK) Shared() bool {
 }
 
 // The NUMA node. The maximum value of this member is 63.
-func (b PSAPI_WORKING_SET_EX_BLOCK) Node() int {
+func (b PSAPI_WORKING_SET_EX_BLOCK) Node() uint64 {
 	return b.intField(16, 6)
 }
 
@@ -1754,14 +1754,14 @@ func (b PSAPI_WORKING_SET_EX_BLOCK) Bad() bool {
 }
 
 // intField extracts an integer field in the PSAPI_WORKING_SET_EX_BLOCK union.
-func (b PSAPI_WORKING_SET_EX_BLOCK) intField(start, length int) int {
+func (b PSAPI_WORKING_SET_EX_BLOCK) intField(start, length int) uint64 {
 	var mask PSAPI_WORKING_SET_EX_BLOCK
 	for pos := start; pos < start+length; pos++ {
 		mask |= (1 << pos)
 	}
 
 	masked := b & mask
-	return int(masked >> start)
+	return uint64(masked >> start)
 }
 
 // PSAPI_WORKING_SET_EX_INFORMATION contains extended working set information for a process.
